move more logic into core
